Extensibility: Refactor hooks to make them reusable for editor and components
