+++ blocks/api/serializer.js
@@ -8,12 +8,12 @@ import { html as beautifyHtml } from 'js-beautify';
  * WordPress dependencies
  */
 import { Component, createElement, renderToString, cloneElement, Children } from '@wordpress/element';
+import { applyFilters } from '@wordpress/utils';
 
 /**
  * Internal dependencies
  */
 import { getBlockType, getUnknownTypeHandlerName } from './registration';
-import { applyFilters } from '../hooks';
 
 /**
  * Returns the block's default classname from its name
