+++ blocks/api/test/factory.js
@@ -21,6 +21,11 @@ describe( 'block factory', () => {
 		title: 'block title',
 	};
 
+	beforeAll( () => {
+		// Load all hooks that modify blocks
+		require( 'blocks/hooks' );
+	} );
+
 	afterEach( () => {
 		setUnknownTypeHandlerName( undefined );
 		getBlockTypes().forEach( ( block ) => {
