+++ blocks/block-edit/index.js
@@ -1,8 +1,12 @@
 /**
+ * WordPress dependencies
+ */
+import { applyFilters } from '@wordpress/utils';
+
+/**
  * Internal dependencies
  */
 import { getBlockType } from '../api';
-import { applyFilters } from '../hooks';
 
 function BlockEdit( props ) {
 	const { name, ...editProps } = props;
@@ -15,10 +19,7 @@ function BlockEdit( props ) {
 	// `edit` and `save` are functions or components describing the markup
 	// with which a block is displayed. If `blockType` is valid, assign
 	// them preferencially as the render value for the block.
-	let Edit;
-	if ( blockType ) {
-		Edit = blockType.edit || blockType.save;
-	}
+	const Edit = blockType.edit || blockType.save;
 
 	return applyFilters( 'BlockEdit', <Edit key="edit" { ...editProps } />, props );
 }
