+++ components/higher-order/with-filters/index.js
@@ -1,20 +1,17 @@
 /**
  * WordPress dependencies
  */
-import createHooks from '@wordpress/hooks';
+import { applyFilters } from '@wordpress/utils';
 
 /**
  * Internal dependencies
  */
 import wrapperDisplayName from '../wrapper-display-name';
 
-// TODO: Use withHooks HOC from https://github.com/WordPress/gutenberg/pull/3321.
-const hooks = createHooks();
-
 export default function withFilters( hookName ) {
 	return ( WrappedComponent ) => {
 		const FiltersComponent = ( props ) => {
-			return hooks.applyFilters( hookName, <WrappedComponent { ...props } />, props );
+			return applyFilters( hookName, <WrappedComponent { ...props } />, props );
 		};
 		FiltersComponent.displayName = wrapperDisplayName( 'filters', WrappedComponent );
