+++ utils/index.js
@@ -8,6 +8,7 @@ export { keycodes };
 export { decodeEntities };
 
 export * from './blob-cache';
+export * from './hooks';
 export * from './mediaupload';
 
 export { viewPort };
